backend/requirements.txt
# requirements.txt
fastapi
httpx
pydantic
uvicorn
python-dotenv
